Revert "Merge remote-tracking branch 'origin/feature/cloud_faas' into feature/cloud_faas"
This reverts commit 022a78288d0d2064b1ea3179c368f74bdf3c8052.
