+++ README.md
@@ -145,15 +145,12 @@ Here's a list of organizations that we know about from our community. Don't hesi
 
 -   [Airbus](https://www.airbus.com/en)
 -   [Alcatel-Lucent](https://www.al-enterprise.com/)
--   [Apple](https://www.apple.com/)
 -   [BetterUp](https://www.betterup.com/)
 -   [Databricks](https://www.databricks.com/)
 -   [Deepset](https://deepset.ai/)
 -   [Etalab](https://www.etalab.gouv.fr/)
 -   [Infineon](https://www.infineon.com/)
--   [Intelijus](https://www.intelijus.ai/)
 -   [LEGO](https://www.lego.com/)
--   [Netflix](https://netflix.com)
 -   [Nvidia](https://www.nvidia.com/en-us/)
 -   [Rakuten](https://www.rakuten.com/)
 -   [Sooth.ai](https://sooth.ai/)
