+++ e2e/conftest.py
@@ -21,15 +21,13 @@ from haystack.document_stores import (
 )
 
 
+SAMPLES_PATH = Path(__file__).parent / "samples"
+
+
 # Fix all random seeds that come to mind
 torch.manual_seed(0)
 np.random.seed(0)
 random.seed(0)
-
-
-@pytest.fixture
-def samples_path():
-    return Path(__file__).parent / "samples"
 
 
 @pytest.fixture
