+++ haystack/nodes/prompt/__init__.py
@@ -2,3 +2,4 @@ from haystack.nodes.prompt.prompt_node import PromptNode
 from haystack.nodes.prompt.prompt_template import PromptTemplate
 from haystack.nodes.prompt.prompt_model import PromptModel
 from haystack.nodes.prompt.shapers import BaseOutputParser, AnswerParser
+from haystack.nodes.prompt.providers import PromptModelInvocationLayer
