--- haystack/nodes/prompt/invocation_layer/__init__.py
@@ -1,6 +1,0 @@
-from haystack.nodes.prompt.invocation_layer.azure_open_ai import AzureOpenAIInvocationLayer
-from haystack.nodes.prompt.invocation_layer.base import PromptModelInvocationLayer
-from haystack.nodes.prompt.invocation_layer.chatgpt import ChatGPTInvocationLayer
-from haystack.nodes.prompt.invocation_layer.handlers import TokenStreamingHandler, DefaultTokenStreamingHandler
-from haystack.nodes.prompt.invocation_layer.hugging_face import HFLocalInvocationLayer
-from haystack.nodes.prompt.invocation_layer.open_ai import OpenAIInvocationLayer
