+++ test/modeling/test_processor_save_load.py
@@ -7,8 +7,10 @@ from haystack.modeling.data_handler.processor import SquadProcessor
 from haystack.modeling.utils import set_all_seeds
 import torch
 
+from ..conftest import SAMPLES_PATH
 
-def test_processor_saving_loading(tmp_path, caplog, samples_path):
+
+def test_processor_saving_loading(tmp_path, caplog):
     if caplog is not None:
         caplog.set_level(logging.CRITICAL)
 
@@ -24,17 +26,17 @@ def test_processor_saving_loading(tmp_path, caplog, samples_path):
         train_filename="train-sample.json",
         dev_filename="dev-sample.json",
         test_filename=None,
-        data_dir=samples_path / "qa",
+        data_dir=SAMPLES_PATH / "qa",
     )
 
-    dicts = processor.file_to_dicts(file=samples_path / "qa" / "dev-sample.json")
+    dicts = processor.file_to_dicts(file=SAMPLES_PATH / "qa" / "dev-sample.json")
     data, tensor_names, _ = processor.dataset_from_dicts(dicts=dicts, indices=[1])
 
     save_dir = tmp_path / Path("testsave/processor")
     processor.save(save_dir)
 
     processor = processor.load_from_dir(save_dir)
-    dicts = processor.file_to_dicts(file=samples_path / "qa" / "dev-sample.json")
+    dicts = processor.file_to_dicts(file=SAMPLES_PATH / "qa" / "dev-sample.json")
     data_loaded, tensor_names_loaded, _ = processor.dataset_from_dicts(dicts, indices=[1])
 
     assert tensor_names == tensor_names_loaded
