+++ test/nodes/test_retriever.py
@@ -27,7 +27,7 @@ from haystack.nodes.retriever.dense import DensePassageRetriever, EmbeddingRetri
 from haystack.nodes.retriever.sparse import BM25Retriever, FilterRetriever, TfidfRetriever
 from haystack.nodes.retriever.multimodal import MultiModalRetriever
 
-from ..conftest import MockRetriever, fail_at_version
+from ..conftest import SAMPLES_PATH, MockRetriever, fail_at_version
 
 
 # TODO check if we this works with only "memory" arg
@@ -657,7 +657,7 @@ def test_table_text_retriever_saving_and_loading(tmp_path, retriever, document_s
 
 
 @pytest.mark.embedding_dim(128)
-def test_table_text_retriever_training(tmp_path, document_store, samples_path):
+def test_table_text_retriever_training(tmp_path, document_store):
     retriever = TableTextRetriever(
         document_store=document_store,
         query_embedding_model="deepset/bert-small-mm_retrieval-question_encoder",
@@ -667,7 +667,7 @@ def test_table_text_retriever_training(tmp_path, document_store, samples_path):
     )
 
     retriever.train(
-        data_dir=samples_path / "mmr",
+        data_dir=SAMPLES_PATH / "mmr",
         train_filename="sample.json",
         n_epochs=1,
         n_gpu=0,
@@ -1044,10 +1044,10 @@ def table_docs() -> List[Document]:
 
 
 @pytest.fixture
-def image_docs(samples_path) -> List[Document]:
+def image_docs() -> List[Document]:
     return [
-        Document(content=str(samples_path / "images" / imagefile), content_type="image")
-        for imagefile in os.listdir(samples_path / "images")
+        Document(content=str(SAMPLES_PATH / "images" / imagefile), content_type="image")
+        for imagefile in os.listdir(SAMPLES_PATH / "images")
     ]
 
 
@@ -1117,7 +1117,7 @@ def test_multimodal_retriever_query():
 
 
 @pytest.mark.integration
-def test_multimodal_image_retrieval(image_docs: List[Document], samples_path):
+def test_multimodal_image_retrieval(image_docs: List[Document]):
     retriever = MultiModalRetriever(
         document_store=InMemoryDocumentStore(return_embedding=True, embedding_dim=512),
         query_embedding_model="sentence-transformers/clip-ViT-B-32",
@@ -1127,12 +1127,12 @@ def test_multimodal_image_retrieval(image_docs: List[Document], samples_path):
     retriever.document_store.update_embeddings(retriever=retriever)
 
     results = retriever.retrieve(query="What's a cat?")
-    assert str(results[0].content) == str(samples_path / "images" / "cat.jpg")
+    assert str(results[0].content) == str(SAMPLES_PATH / "images" / "cat.jpg")
 
 
 @pytest.mark.skip("Not working yet as intended")
 @pytest.mark.integration
-def test_multimodal_text_image_retrieval(text_docs: List[Document], image_docs: List[Document], samples_path):
+def test_multimodal_text_image_retrieval(text_docs: List[Document], image_docs: List[Document]):
     retriever = MultiModalRetriever(
         document_store=InMemoryDocumentStore(return_embedding=True, embedding_dim=512),
         query_embedding_model="sentence-transformers/clip-ViT-B-32",
@@ -1150,7 +1150,7 @@ def test_multimodal_text_image_retrieval(text_docs: List[Document], image_docs: 
     text_results = [result for result in results if result.content_type == "text"]
     image_results = [result for result in results if result.content_type == "image"]
 
-    assert str(image_results[0].content) == str(samples_path / "images" / "paris.jpg")
+    assert str(image_results[0].content) == str(SAMPLES_PATH / "images" / "paris.jpg")
     assert text_results[0].content == "My name is Christelle and I live in Paris"
