+++ test/nodes/test_sampler.py
@@ -35,7 +35,7 @@ docs = [
 ]
 
 
-@pytest.mark.integration
+@pytest.mark.unit
 def test_top_p_sampling(top_p_sampler):
     query = "What is the most important religious learning from the Bible?"
     results = top_p_sampler.predict(query=query, documents=docs, top_p=0.98)
@@ -44,7 +44,7 @@ def test_top_p_sampling(top_p_sampler):
     assert results[1].id == "5"
 
 
-@pytest.mark.integration
+@pytest.mark.unit
 def test_top_p_sampling_top_p_none(top_p_sampler):
     query = "What is the most important religious learning from the Bible?"
     results = top_p_sampler.predict(query=query, documents=docs, top_p=None)
@@ -55,7 +55,7 @@ def test_top_p_sampling_top_p_none(top_p_sampler):
     assert len(results) == 5
 
 
-@pytest.mark.integration
+@pytest.mark.unit
 def test_top_p_sampling_at_least_one_result(top_p_sampler):
     query = "What is the most important building in King's Landing that has a religious background?"
     results = top_p_sampler.predict(query=query, documents=docs, top_p=0.9)
