+++ test/others/test_schema.py
@@ -3,6 +3,8 @@ import pytest
 import numpy as np
 import pandas as pd
 
+from ..conftest import SAMPLES_PATH, fail_at_version
+
 LABELS = [
     Label(
         query="some",
