+++ test/pipelines/test_eval_batch.py
@@ -20,6 +20,8 @@ from haystack.pipelines.standard_pipelines import (
 from haystack.nodes.translator.transformers import TransformersTranslator
 from haystack.schema import Answer, Document, EvaluationResult, Label, MultiLabel, Span
 
+from ..conftest import SAMPLES_PATH
+
 
 @pytest.mark.skipif(sys.platform in ["win32", "cygwin"], reason="Causes OOM on windows github runner")
 @pytest.mark.parametrize("document_store_with_docs", ["memory"], indirect=True)
