+++ test/pipelines/test_ray.py
@@ -5,6 +5,8 @@ import ray
 
 from haystack.pipelines import RayPipeline
 
+from ..conftest import SAMPLES_PATH
+
 
 @pytest.fixture(autouse=True)
 def shutdown_ray():
@@ -21,9 +23,9 @@ def shutdown_ray():
 @pytest.mark.integration
 @pytest.mark.parametrize("document_store_with_docs", ["elasticsearch"], indirect=True)
 @pytest.mark.parametrize("serve_detached", [True, False])
-def test_load_pipeline(document_store_with_docs, serve_detached, samples_path):
+def test_load_pipeline(document_store_with_docs, serve_detached):
     pipeline = RayPipeline.load_from_yaml(
-        samples_path / "pipeline" / "ray.simple.haystack-pipeline.yml",
+        SAMPLES_PATH / "pipeline" / "ray.simple.haystack-pipeline.yml",
         pipeline_name="ray_query_pipeline",
         ray_args={"num_cpus": 8},
         serve_args={"detached": serve_detached},
@@ -43,9 +45,9 @@ def test_load_pipeline(document_store_with_docs, serve_detached, samples_path):
 
 @pytest.mark.integration
 @pytest.mark.parametrize("document_store_with_docs", ["elasticsearch"], indirect=True)
-def test_load_advanced_pipeline(document_store_with_docs, samples_path):
+def test_load_advanced_pipeline(document_store_with_docs):
     pipeline = RayPipeline.load_from_yaml(
-        samples_path / "pipeline" / "ray.advanced.haystack-pipeline.yml",
+        SAMPLES_PATH / "pipeline" / "ray.advanced.haystack-pipeline.yml",
         pipeline_name="ray_query_pipeline",
         ray_args={"num_cpus": 8},
         serve_args={"detached": True},
@@ -71,9 +73,9 @@ def test_load_advanced_pipeline(document_store_with_docs, samples_path):
 @pytest.mark.asyncio
 @pytest.mark.integration
 @pytest.mark.parametrize("document_store_with_docs", ["elasticsearch"], indirect=True)
-async def test_load_advanced_pipeline_async(document_store_with_docs, samples_path):
+async def test_load_advanced_pipeline_async(document_store_with_docs):
     pipeline = RayPipeline.load_from_yaml(
-        samples_path / "pipeline" / "ray.advanced.haystack-pipeline.yml",
+        SAMPLES_PATH / "pipeline" / "ray.advanced.haystack-pipeline.yml",
         pipeline_name="ray_query_pipeline",
         ray_args={"num_cpus": 8},
         serve_args={"detached": True},
