+++ test/samples/crawler/page1.html
@@ -7,7 +7,5 @@
     <p>page 1 content</p>
     <a href="index.html">link to home</a>
     <a href="page2.html">link to page 2</a>
-    <a href="page1_subpage1.html">link to subpage 1</a>
-    <a href="page1_subpage2.html">link to subpage 2</a>
 </body>
 </html>
