--- test/samples/crawler/page1_subpage1.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Test Subpage 1 of Page 1 for Crawler</title>
-</head>
-<body>
-    <p>subpage 1 of page 1 content</p>
-</body>
-</html>
