+++ test/samples/crawler/page2.html
@@ -7,6 +7,5 @@
     <p>page 2 content</p>
     <a href="index.html">link to home</a>
     <a href="page1.html">link to page 1</a>
-    <a href="page2_subpage1.html">link to subpage</a>
 </body>
 </html>
