--- test/samples/crawler/page2_subpage1.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Test Subpage 1 of Page 2 for Crawler</title>
-</head>
-<body>
-    <p>subpage 1 of page 2 content</p>
-</body>
-</html>
